+++ src/renderer_d3d11.cpp
@@ -2902,8 +2902,8 @@ namespace bgfx { namespace d3d11
 				desc.DepthWriteMask = !!(BGFX_STATE_WRITE_Z & _state) ? D3D11_DEPTH_WRITE_MASK_ALL : D3D11_DEPTH_WRITE_MASK_ZERO;
 				desc.DepthFunc      = s_cmpFunc[func];
 
-				uint32_t bstencil     = _bstencil;
-				uint32_t frontAndBack = bstencil != BGFX_STENCIL_NONE && bstencil != fstencil;
+				uint64_t bstencil     = _bstencil;
+				uint64_t frontAndBack = bstencil != BGFX_STENCIL_NONE && bstencil != fstencil;
 				bstencil = frontAndBack ? bstencil : fstencil;
 
 				desc.StencilEnable    = 0 != _stencil;
@@ -3368,7 +3368,7 @@ namespace bgfx { namespace d3d11
 
 			for (;;)
 			{
-				uint32_t opcode = _uniformBuffer.read();
+				uint64_t opcode = _uniformBuffer.read();
 
 				if (UniformType::End == opcode)
 				{
@@ -4143,7 +4143,7 @@ namespace bgfx { namespace d3d11
 				uint8_t type = 0;
 				bx::read(&reader, type, &err);
 
-				uint8_t num = 0;
+				uint16_t num = 0;
 				bx::read(&reader, num, &err);
 
 				uint16_t regIndex = 0;
